feat: user api 파일 생성 후 관심사 api 추가
